Refactoring the term classes
The information of whether or not an index is a creator/annihilator is
now stored directly inside the Index instead of indices being stored in
different lists inside the Tensor class.

Furthermore an Index now contains information about whether or not is is
spin-affiliated (meaning that it implicitly runs over alpha and beta
spins which effectively doubles its range).