Use multiprecision integer type for contraction cost
When contracting only 8 indices of which four belong to a space of size
10 and the other four to a space of size 100, we already have a cost of
1E12. That is beyond the range of a 32-bit number. With 64-bit we get up
to 1E19 but it can be seen that this limit can be easily broken when
calculating these kind of things.

Thus the type for representing conraction costs has been changed to be a
multiprecision integer type (from Boost). It is now set to be 512-bit
which should be sufficient for a while.

It was also considered to use logarithmic values for the costs but due
to the log function being a somewhat expensive function, I figured the
multiprecision approach might be faster. Plus the log approach had other
issues as well (e.g. how to represent 0, how to add costs, etc)