Also apply factor factor to fully symmetric terms
This changes the way the prefactor for terms is determined. The new
implementation no longer checks whether a prefactor is needed for the
initial antisymmetrization since we only antisymmetrize if it is indeed
necessary which in turn means that we don't want to apply a factor for
that anway.

However we now add a factor for fully particle-1,2-symmetric terms,
since this symmetrization is performed "blindly" in the end. The
prefactor makes sure that this does not change anything for these terms.