Fix bug in comparator for PermutationGroup::Element
Including the factor in the comparison made it almost impossible for an
element to become smaller than another one in a few circumstances. Since
we don't include the factor in the check for equality since we don't
expect the same sequence to be reached with different factors (or
multiple times at all), we also don't have to check for it in the
less-than comparison.