Fix issue with differing symmetries
Before explicit symmetrization of the end-result(s), it could happen
that different terms produced "asymmetric" end-results (*-u) that
weren't actually asymmetric. That would lead to different versions of
these *-u tensors that only differ in their symmetry.

However since ultimately we are considering all of them to be the same
tensor, we have to make sure to only pick the least symmetric one, since
that's the one that determines the overall symmetry of the
to-be-symmetrized end-result.