Make spin-integration symmetry-independent
The previous implementation had the assumption that in order for certain
spin-cases to exist, a tensor must be at least partially antisymmetric.
As it turns out though, this is not the case. This only applies to
special tensors such as the amplitudes or the 2-electron-integrals.
However since we assume to always be using the symmetrized versions of
those, this symmetry-restriction was completely removed from the code in
order to account for all possible spin-cases.