Add fallback for older Boost versions
We require the is_detected type_trait which was only introduced in Boost
1.67.0. However, since the type_traits library is more or less
self-contained and does not have big dependencies on other Boost
libraries, this commit adds a workaround that will simply download the
Boost type_traits library in version 1.67.0 and add it to the
include-path for the program. That way the is_detected header should be
found and be usable, even if the overall Boost version that is used does
not contain it.