BUILD: Only explicitly link to fs, if required
If the code compiles fine, without explicitly linking to the
std-filesystem library (as is the case for more modern compilers), we
don't do it. This has the advantage of not breaking builds that would
work without the explicit linking but break when doing so (this is the
case on e.g. macOS).